backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
pydantic
python-multipart
openai
python-docx
pypdf
pdfplumber
pytest
pytest-cov
